hotshot_query_service::availability
Module query_data

Structs

BlockQueryData

BlockSummaryQueryData

HeaderQueryData

InconsistentLeafError

InconsistentLeafSnafu 🔒

SNAFU context selector for the InconsistentLeafError error

LeafQueryData

Limits

PayloadMetadata

A summary of a payload without all the data.

PayloadQueryData

TransactionQueryData

TransactionSummaryQueryData

VidCommonMetadata

A summary of a VID payload without all the data.

VidCommonQueryData

Traits

QueryableHeader

QueryablePayload

A block payload whose contents (e.g. individual transactions) can be examined.

Functions

payload_size 🔒

Type Aliases

BlockHash

A block hash is the hash of the block header.

LeafHash

QcHash

Timestamp

TransactionHash

TransactionInclusionProof

TransactionIndex