hotshot_query_service::data_source::fetching

Trait Storable

trait Storable<Types: NodeType>: Clone {
    // Required methods
    fn debug_name(&self) -> String;
    fn notify(
        &self,
        notifiers: &Notifiers<Types>,
    ) -> impl Send + Future<Output = ()>;
    fn store(
        &self,
        storage: &mut impl UpdateAvailabilityStorage<Types>,
        leaf_only: bool,
    ) -> impl Send + Future<Output = Result<()>>;
}

An object which can be stored in the database.

Required Methods

fn debug_name(&self) -> String

The name of this object, for debugging purposes.

fn notify( &self, notifiers: &Notifiers<Types>, ) -> impl Send + Future<Output = ()>

Notify anyone waiting for this object that it has become available.

fn store( &self, storage: &mut impl UpdateAvailabilityStorage<Types>, leaf_only: bool, ) -> impl Send + Future<Output = Result<()>>

Store the object in the local database.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.