hotshot_query_service::data_source::notifier
Trait Predicate

pub trait Predicate<T>:
    'static
    + Send
    + Sync
    + Fn(&T) -> bool { }

A predicate on a type <T>.

Predicate is an alias for any type implementing Fn(&T) -> bool (with a few extra bounds to support concurrency). It is used by Notifier to represent the preferences of subscribers when filtering messages of type T.