Module hotshot_query_service::data_source::sql

Re-exports

pub use sql::Config;

pub use sql::Query;

pub use sql::Transaction;

pub use tokio_postgres as postgres;

Modules

testing

Macros

include_migrations

Embed migrations from the given directory into the current binary.

Structs

Error

The Error type, a wrapper around a dynamic error type.

Migration

Represents a schema migration to be run on the database, this struct is used by the embed_migrations! macro to gather migration files and shouldn’t be needed by the user

Type Aliases

Builder

SqlDataSource

A data source for the APIs provided in this crate, backed by a remote PostgreSQL database.