Module hotshot_query_service::data_source::storage

Persistent storage for data sources.

Naturally, an archival query service such as this is heavily dependent on a persistent storage implementation. This module defines the interfaces required of this storage. Any storage layer implementing the appropriate interfaces can be used as the storage layer when constructing a FetchingDataSource, which can in turn be used to instantiate the REST APIs provided by this crate.

This module also comes with a few pre-built persistence implementations:

SqlStorage

FileSystemStorage

NoStorage

Re-exports

pub use fs::FileSystemStorage;

pub use no_storage::NoStorage;

pub use sql::SqlStorage;

Modules

fs

no_storage

pruning

sql

Traits

AvailabilityStorage

Persistent storage for a HotShot blockchain.

ExplorerStorage

An interface for querying Data and Statistics from the HotShot Blockchain.