hotshot_query_service::data_source::storage
Module fs

Structs

FileSystemStorage

Storage for the APIs provided in this crate, backed by a remote PostgreSQL database.

FileSystemStorageInner

Transaction

Constants

CACHED_BLOCKS_COUNT 🔒

CACHED_LEAVES_COUNT 🔒

CACHED_VID_COMMON_COUNT 🔒

Traits

Revert

Functions

range_iter 🔒

update_index_by_hash 🔒

Update an index mapping hashes of objects to their positions in the ledger.