hotshot_query_service::data_source::storage::sql
Function add_custom_migrations

fn add_custom_migrations(
    default: impl IntoIterator<Item = Migration>,
    custom: impl IntoIterator<Item = Migration>,
) -> impl Iterator<Item = Migration>

Add custom migrations to a default migration sequence.

Migrations in custom replace migrations in default with the same version. Otherwise, the two sequences default and custom are merged so that the resulting sequence is sorted by ascending version number. Each of default and custom is assumed to be the output of validate_migrations; that is, each is sorted by version and contains no duplicate versions.