hotshot_query_service::data_source::storage::sql::queries
Module explorer

Explorer storage implementation for a database query engine.

Structs

GET_BLOCK_DETAIL_QUERY_FOR_HASH 🔒

GET_BLOCK_DETAIL_QUERY_FOR_HEIGHT 🔒

GET_BLOCK_DETAIL_QUERY_FOR_LATEST 🔒

GET_BLOCK_SUMMARIES_QUERY_FOR_HASH 🔒

GET_BLOCK_SUMMARIES_QUERY_FOR_HEIGHT 🔒

GET_BLOCK_SUMMARIES_QUERY_FOR_LATEST 🔒

GET_TRANSACTION_DETAIL_QUERY_FOR_HASH 🔒

GET_TRANSACTION_DETAIL_QUERY_FOR_HEIGHT_AND_OFFSET 🔒

GET_TRANSACTION_DETAIL_QUERY_FOR_LATEST 🔒

GET_TRANSACTION_SUMMARIES_QUERY_FOR_BLOCK 🔒

GET_TRANSACTION_SUMMARIES_QUERY_FOR_NO_FILTER 🔒

Constants

EXPLORER_SUMMARY_HISTOGRAM_NUM_ENTRIES 🔒

EXPLORER_SUMMARY_HISTOGRAM_NUM_ENTRIES is the number of entries we want to return in our histogram summary.

EXPLORER_SUMMARY_NUM_BLOCKS 🔒

EXPLORER_SUMMARY_NUM_BLOCKS is the number of blocks we want to return in our explorer summary.

EXPLORER_SUMMARY_NUM_TRANSACTIONS 🔒

EXPLORER_SUMMARY_NUM_TRANSACTIONS is the number of transactions we want to return in our explorer summary.