hotshot_query_service::data_source::storage::sql
Struct Config

pub struct Config {
    db_opt: PgConnectOptions,
    pool_opt: PoolOptions<Db>,
    schema: String,
    reset: bool,
    migrations: Vec<Migration>,
    no_migrations: bool,
    pruner_cfg: Option<PrunerCfg>,
    archive: bool,
    pool: Option<Pool<Db>>,
}

Fields
db_opt: PgConnectOptionspool_opt: PoolOptions<Db>schema: Stringreset: boolmigrations: Vec<Migration>no_migrations: boolpruner_cfg: Option<PrunerCfg>archive: boolpool: Option<Pool<Db>>
Implementations

impl Config

pub async fn connect<Types, P: AvailabilityProvider<Types>>( self, provider: P, ) -> Result<SqlDataSource<Types, P>, Error>
where Types: NodeType, Header<Types>: QueryableHeader<Types>, Payload<Types>: QueryablePayload<Types>,

Connect to the database with this config.

pub async fn builder<Types, P: AvailabilityProvider<Types>>( self, provider: P, ) -> Result<Builder<Types, P>, Error>
where Types: NodeType, Header<Types>: QueryableHeader<Types>, Payload<Types>: QueryablePayload<Types>,

Connect to the database, setting options on the underlying FetchingDataSource using the fetching::Builder interface.

impl Config

pub fn host(self, host: impl Into<String>) -> Self

Set the hostname of the database server.

The default is localhost.

pub fn port(self, port: u16) -> Self

Set the port on which to connect to the database.

The default is 5432, the default Postgres port.

pub fn user(self, user: &str) -> Self

Set the DB user to connect as.

pub fn password(self, password: &str) -> Self

Set a password for connecting to the database.

pub fn database(self, database: &str) -> Self

Set the name of the database to connect to.

pub fn tls(self) -> Self

Use TLS for an encrypted connection to the database.

Note that an encrypted connection may be established even if this option is not set, as long as both the client and server support it. This option merely causes connection to fail if an encrypted stream cannot be established.

pub fn schema(self, schema: impl Into<String>) -> Self

Set the name of the schema to use for queries.

The default schema is named hotshot and is created via the default migrations.

impl Config

pub fn pool(self, pool: Pool<Db>) -> Self

Sets the database connection pool This allows reusing an existing connection pool when building a new SqlStorage instance.

pub fn reset_schema(self) -> Self

Reset the schema on connection.

When this Config is used to connect a SqlDataSource, if this option is set, the relevant schema will first be dropped and then recreated, yielding a completely fresh instance of the query service.

This is a particularly useful capability for development and staging environments. Still, it must be used with extreme caution, as using this will irrevocably delete any data pertaining to the query service in the database.

pub fn migrations(self, migrations: impl IntoIterator<Item = Migration>) -> Self

Add custom migrations to run when connecting to the database.

pub fn no_migrations(self) -> Self

Skip all migrations when connecting to the database.

pub fn pruner_cfg(self, cfg: PrunerCfg) -> Result<Self, Error>

Enable pruning with a given configuration.

If archive was previously specified, this will override it.

pub fn archive(self) -> Self

Disable pruning and reconstruct previously pruned data.

While running without pruning is the default behavior, the default will not try to reconstruct data that was pruned in a previous run where pruning was enabled. This option instructs the service to run without pruning and reconstruct all previously pruned data by fetching from peers.

If pruner_cfg was previously specified, this will override it.

pub fn idle_connection_timeout(self, timeout: Duration) -> Self

Set the maximum idle time of a connection.

Any connection which has been open and unused longer than this duration will be automatically closed to reduce load on the server.

pub fn connection_timeout(self, timeout: Duration) -> Self

Set the maximum lifetime of a connection.

Any connection which has been open longer than this duration will be automatically closed (and, if needed, replaced), even if it is otherwise healthy. It is good practice to refresh even healthy connections once in a while (e.g. daily) in case of resource leaks in the server implementation.

pub fn min_connections(self, min: u32) -> Self

Set the minimum number of connections to maintain at any time.

The data source will, to the best of its ability, maintain at least min open connections at all times. This can be used to reduce the latency hit of opening new connections when at least this many simultaneous connections are frequently needed.

pub fn max_connections(self, max: u32) -> Self

Set the maximum number of connections to maintain at any time.

Once max connections are in use simultaneously, further attempts to acquire a connection (or begin a transaction) will block until one of the existing connections is released.

pub fn slow_statement_threshold(self, threshold: Duration) -> Self

Log at WARN level any time a SQL statement takes longer than threshold.

The default threshold is 1s.