hotshot_query_service::data_source::storage::sql
Trait FixedLengthParams

pub trait FixedLengthParams<'p, const N: usize>: Params<'p> { }

A collection of parameters with a statically known length.

This is a simple trick for enforcing at compile time that a list of parameters has a certain length, such as matching the length of a list of column names. This can prevent easy mistakes like leaving out a parameter. It is implemented for tuples up to length 8.