Trait hotshot_query_service::data_source::storage::sql::Query

pub trait Query {
    // Required method
    fn client<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Cow<'_, Arc<Client>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn query<'life0, 'life1, 'async_trait, T, P>(
        &'life0 self,
        query: &'life1 T,
        params: P
    ) -> Pin<Box<dyn Future<Output = QueryResult<BoxStream<'static, QueryResult<Row>>>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             P: IntoIterator + Send + 'async_trait,
             P::IntoIter: ExactSizeIterator,
             P::Item: BorrowToSql,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn query_static<'life0, 'life1, 'async_trait, T>(
        &'life0 self,
        query: &'life1 T
    ) -> Pin<Box<dyn Future<Output = QueryResult<BoxStream<'static, QueryResult<Row>>>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn query_one<'life0, 'life1, 'async_trait, T, P>(
        &'life0 self,
        query: &'life1 T,
        params: P
    ) -> Pin<Box<dyn Future<Output = QueryResult<Row>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             P: IntoIterator + Send + 'async_trait,
             P::IntoIter: ExactSizeIterator,
             P::Item: BorrowToSql,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn query_one_static<'life0, 'life1, 'async_trait, T>(
        &'life0 self,
        query: &'life1 T
    ) -> Pin<Box<dyn Future<Output = QueryResult<Row>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn query_opt<'life0, 'life1, 'async_trait, T, P>(
        &'life0 self,
        query: &'life1 T,
        params: P
    ) -> Pin<Box<dyn Future<Output = QueryResult<Option<Row>>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             P: IntoIterator + Send + 'async_trait,
             P::IntoIter: ExactSizeIterator,
             P::Item: BorrowToSql,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn query_opt_static<'life0, 'life1, 'async_trait, T>(
        &'life0 self,
        query: &'life1 T
    ) -> Pin<Box<dyn Future<Output = QueryResult<Option<Row>>> + Send + 'async_trait>>
       where T: ?Sized + ToStatement + Sync + 'async_trait,
             Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Required Methods

fn client<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Cow<'_, Arc<Client>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn query<'life0, 'life1, 'async_trait, T, P>( &'life0 self, query: &'life1 T, params: P ) -> Pin<Box<dyn Future<Output = QueryResult<BoxStream<'static, QueryResult<Row>>>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, P: IntoIterator + Send + 'async_trait, P::IntoIter: ExactSizeIterator, P::Item: BorrowToSql, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn query_static<'life0, 'life1, 'async_trait, T>( &'life0 self, query: &'life1 T ) -> Pin<Box<dyn Future<Output = QueryResult<BoxStream<'static, QueryResult<Row>>>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the underlying SQL database with no parameters.

fn query_one<'life0, 'life1, 'async_trait, T, P>( &'life0 self, query: &'life1 T, params: P ) -> Pin<Box<dyn Future<Output = QueryResult<Row>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, P: IntoIterator + Send + 'async_trait, P::IntoIter: ExactSizeIterator, P::Item: BorrowToSql, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the underlying SQL database, returning exactly one result or failing.

fn query_one_static<'life0, 'life1, 'async_trait, T>( &'life0 self, query: &'life1 T ) -> Pin<Box<dyn Future<Output = QueryResult<Row>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the underlying SQL database with no parameters, returning exactly one result or failing.

fn query_opt<'life0, 'life1, 'async_trait, T, P>( &'life0 self, query: &'life1 T, params: P ) -> Pin<Box<dyn Future<Output = QueryResult<Option<Row>>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, P: IntoIterator + Send + 'async_trait, P::IntoIter: ExactSizeIterator, P::Item: BorrowToSql, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the underlying SQL database, returning zero or one results.

fn query_opt_static<'life0, 'life1, 'async_trait, T>( &'life0 self, query: &'life1 T ) -> Pin<Box<dyn Future<Output = QueryResult<Option<Row>>> + Send + 'async_trait>>
where T: ?Sized + ToStatement + Sync + 'async_trait, Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the underlying SQL database with no parameters, returning zero or one results.

Object Safety

This trait is not object safe.