Struct hotshot_query_service::data_source::QueryData

pub struct QueryData<Types: NodeType, I: NodeImplementation<Types>, UserData>where
    Block<Types>: QueryableBlock,{ /* private fields */ }

Data used by the APIs provided in this crate, including persistent storage.

QueryData is designed to be both extensible (so you can add additional state to the API modules defined in this crate) and composable (so you can use QueryData as one component of a larger state type for an application with additional modules). Extending QueryData is possible through the UserData type parameter – QueryData implements AsRef<UserData> and AsMut<UserData>, so your API extensions can always access UserData from QueryData.

Composing QueryData with other module states is in principle simple – just create an aggregate struct containing both QueryData and your additional module states. A complication arises from how persistent storage is managed: if other modules have their own persistent state, should the storage of QueryData and the other modules be completely independent, or synchronized under the control of a single [AtomicStore]? QueryData supports both patterns: when you create it with create or open, it will open its own [AtomicStore] and manage the synchronization of its own storage, independent of any other persistent data it might be composed with. But when you create it with create_with_store or open_with_store, you may ask it to register its persistent data structures with an existing [AtomicStoreLoader]. If you register other modules’ persistent data structures with the same loader, you can create one [AtomicStore] that synchronizes all the persistent data. Note, though, that when you choose to use create_with_store or open_with_store, you become responsible for ensuring that calls to [AtomicStore::commit_version] alternate with calls to QueryData::commit_version or QueryData::skip_version.

Implementations

impl<Types: NodeType, I: NodeImplementation<Types>, UserData> QueryData<Types, I, UserData>where Block<Types>: QueryableBlock,

pub fn create( path: &Path, user_data: UserData ) -> Result<Self, PersistenceError>

Create a new QueryData with storage at path.

If there is already data at path, it will be archived.

The QueryData will manage its own persistence synchronization.

pub fn open(path: &Path, user_data: UserData) -> Result<Self, PersistenceError>where Deltas<Types, I>: Resolvable<Block<Types>>,

Open an existing QueryData from storage at path.

If there is no data at path, a new store will be created.

The QueryData will manage its own persistence synchronization.

pub fn create_with_store( loader: &mut AtomicStoreLoader, user_data: UserData ) -> Result<Self, PersistenceError>

Create a new QueryData using a persistent storage loader.

If there is existing data corresponding to the QueryData data structures, it will be archived.

The QueryData will register its persistent data structures with loader. The caller is responsible for creating an [AtomicStore] from loader and managing synchronization of the store.

pub fn open_with_store( loader: &mut AtomicStoreLoader, user_data: UserData ) -> Result<Self, PersistenceError>where Deltas<Types, I>: Resolvable<Block<Types>>,

Open an existing QueryData using a persistent storage loader.

If there is no existing data corresponding to the QueryData data structures, a new store will be created.

The QueryData will register its persistent data structures with loader. The caller is responsible for creating an [AtomicStore] from loader and managing synchronization of the store.

pub async fn commit_version(&mut self) -> Result<(), PersistenceError>

Commit the current state to persistent storage.

If the QueryData is managing its own [AtomicStore] (i.e. it was created with create or open) it will update the global version as well. Otherwise, the caller is responsible for calling [AtomicStore::commit_version] after calling this function.

pub fn skip_version(&mut self) -> Result<(), PersistenceError>

Advance the version of the persistent store without committing changes to persistent state.

This function is useful when the [AtomicStore] synchronizing storage for this QueryData is being managed by the caller. The caller may want to persist some changes to other modules whose state is managed by the same [AtomicStore]. In order to call [AtomicStore::commit_version], the version of this QueryData must be advanced, either by commit_version or, if there are no outstanding changes, skip_version.

pub fn revert_version(&mut self) -> Result<(), PersistenceError>

Revert changes made to persistent storage since the last call to commit_version.