Trait hotshot_query_service::data_source::UpdateDataSource

pub trait UpdateDataSource<Types: NodeTypes> {
    type Error: Error + Debug;

    fn metrics(&self) -> Box<dyn Metrics>;
    fn update(&mut self, event: &Event<Types>) -> Result<(), Self::Error>;
}

An extension trait for types which implement the update trait for each API module.

If a type implements both UpdateAvailabilityData and UpdateStatusData, then it can be fully kept up to date through two interfaces:

metrics, to get a handle for populating the status metrics, which should be used when initializing a HotShotHandle

update, to update the query state when a new HotShot event is emitted

Required Associated Types

type Error: Error + Debug

Required Methods

fn metrics(&self) -> Box<dyn Metrics>

Get a handle for populating status metrics.

This function should be called before creating a HotShotHandle, and the resulting [Metrics] handle should be passed to HotShot, which will update it.

fn update(&mut self, event: &Event<Types>) -> Result<(), Self::Error>

Update query state based on a new consensus event.

The caller is responsible for authenticating event. This function does not perform any authentication, and if given an invalid event (one which does not follow from the latest known state of the ledger) it may panic or silently accept the invalid event. This allows the best possible performance in the case where the query service and the HotShot instance are running in the same process (and thus the event stream, directly from HotShot) is trusted.

If you want to update the data source with an untrusted event, for example one received from a peer over the network, you must authenticate it first.