hotshot_query_service::explorer
Module errors

Structs

BadQuery

BadQuery is an error that indicates a submitted query to the service is invalid, or malformed.

InvalidLimit

InvalidLimit is an error that represents that the there was a problem with the given limit parameter.

NotFound

NotFound is an error that represents results could not be found for the given key, or “search” parameters.

QueryError

QueryError is an error that indicates that a specific error occurred while evaluating a query, or decoding the results of a query.

Unimplemented

Unimplemented is an error that indicates that the feature in question is no implemented, but in addition, that this status can be conveyed to the called with serializable error code.

Traits

ExplorerAPIError

ExplorerAPIError is a trait that represents an error that can be returned returned from the ExplorerAPI for various reasons.