hotshot_query_service::explorer::errors
Trait ExplorerAPIError

pub trait ExplorerAPIError: Display + Debug {
    // Required method
    fn code(&self) -> &str;
}

ExplorerAPIError is a trait that represents an error that can be returned returned from the ExplorerAPI for various reasons.

It aims to be Serializable for the purposes of conveying the error to the client.

Required Methods

fn code(&self) -> &str

The code for this error will uniquely identify this specific error.

This value SHOULD match the rename field for the serde tag, if applicable.