hotshot_query_service::explorer
Module query_data

Structs

BlockDetail

BlockDetail is a struct that represents the details of a specific block for use in a Block Explorer.

BlockRange

BlockRange is a struct that represents a range for a specific set of blocks, starting at the given BlockIdentifier.

BlockSummary

BlockSummary is a struct that represents a summary overview of a specific block. It does not have all of the details of a BlockDetail, but it is useful for displaying information in a list of Blocks.

ExplorerHistograms

ExplorerHistograms provides a series of data points that can be used to draw simple histograms for the Block Explorer. The data returned is meant to be an optimal packing of the values being returned.

ExplorerSummary

ExplorerSummary is a struct that represents an at-a-glance snapshot of the Block Chain. It contains some helpful information that can be used to display a simple health check of the Block Chain.

FeeAttribution

FeeAttribution represents a specific attribution of fees for a specific purpose.

GenesisOverview

GenesisOverview provides a summary overview of the block chain since it’s genesis. At a high level it includes the total number of unique rollups, transactions, and blocks that are in the block chain.

GetBlockSummariesRequest

GetBlockSummariesRequest is a struct that represents an incoming request for Block Summaries. This isn’t sent on the line, but an endpoint will be mapped to this struct in order for the request to be processed.

GetTransactionSummariesRequest

GetTransactionSummariesRequest is a struct that represents an incoming request for Transaction Summaries. This isn’t sent on the line, but an endpoint will be mapped to this struct in order for the request to be processed.

SearchResult

SearchResult is a struct that represents the results of executing a search query against the chain. It contains a list of blocks and transactions that match the search query.

Timestamp

Timestamp represents a specific point in time that has a possible offset.

TransactionDetail

TransactionDetail is a struct that represents the details of a specific transaction / payload contained within a Block.

TransactionDetailResponse

TransactionDetailResponse is a struct that represents the information returned concerning a request for a Transaction Detail. It contains the data payloads separately from the details of the Transaction itself.

TransactionRange

TransactionRange is a struct that represents a range for a specific set of transactions, starting at the given TransactionIdentifier.

TransactionSummary

TransactionSummary is a struct that represents a summary overview of a specific transaction / payload contained within a Block. It does not have all of the details of a TransactionDetail, but it is useful for displaying information in a list of Transactions.

Enums

BlockIdentifier

BlockIdentifier is an enum that represents multiple ways of referring to a specific Block. These use cases are specific to a Block Explorer and can be used to reference a specific individual block.

GetBlockDetailError

GetBlockDetailError represents an error that has occurred in response to the get_block_detail request.

GetBlockSummariesError

GetBlockSummariesError represents an error that has occurred in response to the GetBlockSummariesRequest request.

GetExplorerSummaryError

GetExplorerSummaryError represents an error that has occurred in response to the get_explorer_summary request.

GetSearchResultsError

GetSearchResultsError represents an error that has occurred in response to the get_search_results request.

GetTransactionDetailError

GetTransactionDetailError represents an error that has occurred in response to the get_transaction_detail request.

GetTransactionSummariesError

GetTransactionSummariesError represents an error that has occurred in response to the GetTransactionSummariesRequest request.

TimestampConversionError

TimestampConversionError represents an error that has occurred when attempting to convert a timestamp from a specific format to another. It is primarily used when attempting to deserialize a Timestamp from its serialized string representation.

TransactionIdentifier

TransactionIdentifier is an enum that represents multiple ways of of identifying a specific Transaction. These use cases are specific to a Block Explorer and can be used to ideally uniquely identify a Transaction within the Block Chain.

TransactionSummaryFilter

TransactionSummaryFilter represents the various filters that can be applied when retrieving a list of TransactionSummary entries.

Type Aliases

BalanceAmount

BlockNamespaceId

ProposerId

TransactionNamespaceId

WalletAddress