hotshot_query_service::explorer
Trait ExplorerTransaction

pub trait ExplorerTransaction {
    type NamespaceId: Clone + Debug + Serialize + DeserializeOwned + Send + Sync + PartialEq + Eq;

    // Required methods
    fn namespace_id(&self) -> Self::NamespaceId;
    fn payload_size(&self) -> u64;
}

ExplorerTransaction is a trait that allows the Explorer API to be able to retrieve a namespace id from a transaction. This is necessary for the Explorer API to be able to display the namespace id for a TransactionSummary.

Required Associated Types

type NamespaceId: Clone + Debug + Serialize + DeserializeOwned + Send + Sync + PartialEq + Eq

NamespaceId is a type that represents the id of a namespace. It does not indicate how this namespace id is represented, just that there is a representation of it that adheres to the trait restrictions specified.

Required Methods

fn namespace_id(&self) -> Self::NamespaceId

namespace_id returns the namespace id of the individual transaction.

fn payload_size(&self) -> u64

payload_size returns the size of the payload of the transaction.