hotshot_query_service::fetching::provider::any

Trait DebugProvider

trait DebugProvider<Types, T>: Provider<Types, T> + Debug
where
    Types: NodeType,
    T: Request<Types>,
{ }

Blanket trait combining Debug and Provider.

This is necessary to create a fetcher trait object (dyn Provider, see PayloadProvider and LeafProvider) which also implements Debug, since trait objects can only have one non-auto trait bound.