Module hotshot_query_service::merklized_state

Api for querying merklized state

The state API provides an interface for serving queries against arbitrarily old snapshots of the state. This allows a full Merkle tree to be reconstructed from storage. If any parent state is missing then the partial snapshot can not be queried.

Structs

Options

Enums

Error

Snapshot

Traits

MerklizedState

This trait should be implemented by the MerkleTree that the module is initialized for. It defines methods utilized by the module.

MerklizedStateDataSource

This trait defines methods that a data source should implement It enables retrieval of the membership path for a leaf node, which can be used to reconstruct the Merkle tree state.

UpdateStateData

This trait defines methods for updating the storage with the merkle tree state.

UpdateStateStorage

This trait should be implemented by types that represent validated states. It enables services using the module to implement update functionality. Some services may manage multiple Merkle trees within their validated state, requiring updates for all state types with each event. Therefore, the insert_merkle_nodes method of UpdateStateData should be called for all state types within the validated state.

Functions

define_api