hotshot_query_service::merklized_state
Module data_source

Data for the merklized_state API.

This module facilitates storing the state of a Merkle Tree at a specific point in time and provides methods for querying and reconstructing the snapshot.

Enums

Snapshot

Snapshot can be queried by block height (index) or merkle tree commitment

Traits

MerklizedState

This trait should be implemented by the MerkleTree that the module is initialized for. It defines methods utilized by the module.

MerklizedStateDataSource

This trait defines methods that a data source should implement It enables retrieval of the membership path for a leaf node, which can be used to reconstruct the Merkle tree state.

MerklizedStateHeightPersistence

UpdateStateData

This trait defines methods for updating the storage with the merkle tree state.

Type Aliases

StateCommitment 🔒