hotshot_query_service::node
Module data_source

Data for the node API.

This module is just an alternative view of the same data provided by the availability API. It provides more insight into what data the node actually has at present, as opposed to trying to present a perfect view of an abstract chain, fetching data from other sources as needed. It is also more liberal with provided aggregate counts and statistics which may be inaccurate if data is missing.

Due to this relationship with the availability module, this module has its own data source trait but not its own update trait. The node data source is expected to read its data from the same underlying database as the availability API, and as such the data is updated implicitly via the availability API update trait.

Enums

WindowStart

Traits

NodeDataSource