Module hotshot_query_service::status

Queries for node-specific state and uncommitted data.

Unlike the availability and node APIs, which deal only with committed data (albeit with different consistency properties), the status API offers a glimpse into internal consensus state and uncommitted data. Here you can find low-level information about a particular node, such as consensus and networking metrics.

The status API is intended to be a lightweight way to inspect the activities and health of a consensus node. It is the only API that can be run without any persistent storage, and its memory overhead is also very low. As a consequence, it only serves two types of data:

snapshots of the state right now, with no way to query historical snapshots

summary statistics

Structs

Options

Enums

Error

Traits

StatusDataSource

UpdateStatusData

Functions

define_api