Trait hotshot_query_service::testing::consensus::DataSourceLifeCycle

pub trait DataSourceLifeCycle: Send + Sync + Sized + 'static {
    type Storage: Send + Sync;

    // Required methods
    fn create<'async_trait>(
        node_id: usize
    ) -> Pin<Box<dyn Future<Output = Self::Storage> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn connect<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn reset<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_event<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: &'life1 Event<MockTypes>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided method
    fn setup<'life0, 'async_trait>(
        _network: &'life0 mut MockNetwork<Self>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

Required Associated Types

type Storage: Send + Sync

Backing storage for the data source.

This can be used to connect to data sources to the same underlying data. It must be kept alive as long as the related data sources are open.

Required Methods

fn create<'async_trait>( node_id: usize ) -> Pin<Box<dyn Future<Output = Self::Storage> + Send + 'async_trait>>
where Self: 'async_trait,

fn connect<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn reset<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = Self> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn handle_event<'life0, 'life1, 'async_trait>( &'life0 mut self, event: &'life1 Event<MockTypes> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Provided Methods

fn setup<'life0, 'async_trait>( _network: &'life0 mut MockNetwork<Self> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Setup runs after setting up the network but before starting a test.

Object Safety

This trait is not object safe.