hotshot_query_service::testing::mocks
Type Alias MockQuorumProposal

pub type MockQuorumProposal = QuorumProposal<MockTypes>;

Aliased Type

struct MockQuorumProposal {
    pub block_header: TestBlockHeader,
    pub view_number: ViewNumber,
    pub justify_qc: SimpleCertificate<MockTypes, QuorumData<MockTypes>, SuccessThreshold>,
    pub upgrade_certificate: Option<SimpleCertificate<MockTypes, UpgradeProposalData<MockTypes>, UpgradeThreshold>>,
    pub proposal_certificate: Option<ViewChangeEvidence<MockTypes>>,
}

Fields
block_header: TestBlockHeader

The block header to append

view_number: ViewNumber

CurView from leader when proposing leaf

justify_qc: SimpleCertificate<MockTypes, QuorumData<MockTypes>, SuccessThreshold>

Per spec, justification

upgrade_certificate: Option<SimpleCertificate<MockTypes, UpgradeProposalData<MockTypes>, UpgradeThreshold>>

Possible upgrade certificate, which the leader may optionally attach.

proposal_certificate: Option<ViewChangeEvidence<MockTypes>>

Possible timeout or view sync certificate.

A timeout certificate is only present if the justify_qc is not for the preceding view

A view sync certificate is only present if the justify_qc and timeout_cert are not present.