hotshot_query_service
Trait Resolvable

pub trait Resolvable<T: Committable>: Sized {
    // Required methods
    fn try_resolve(self) -> Result<T, Self>;
    fn commitment(&self) -> Commitment<T>;
}

A reference to a T which can be resolved into a whole T.

Required Methods

fn try_resolve(self) -> Result<T, Self>

Get the underlying object if it is available without blocking.

fn commitment(&self) -> Commitment<T>

Get a commitment to the underlying object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.