hotshot_query_service
Module availability

Queries for HotShot chain state.

The availability API provides an objective view of the HotShot blockchain. It provides access only to normative data: that is, data which is agreed upon by all honest consensus nodes and which is immutable. This means access to core consensus data structures including leaves, blocks, and headers, where each query is pure and idempotent. This also means that it is possible for a client to verify all of the information provided by this API, by running a HotShot light client and downloading the appropriate evidence with each query.

This API does not provide any queries which represent only the current state of the chain or may change over time, and it does not provide information for which there is not (yet) agreement of a supermajority of consensus nodes. For information about the current dynamic state of consensus and uncommitted state, try the status API. For information about the chain which is tabulated by this specific node and not subject to full consensus agreement, try the node API.

Modules

data_source 🔒

fetch 🔒

query_data 🔒

Structs

BlockInfo

Information about a block.

BlockQueryData

BlockSummaryQueryData

CustomSnafu

SNAFU context selector for the Error::Custom variant

FetchBlockSnafu

SNAFU context selector for the Error::FetchBlock variant

FetchHeaderSnafu

SNAFU context selector for the Error::FetchHeader variant

FetchLeafSnafu

SNAFU context selector for the Error::FetchLeaf variant

FetchTransactionSnafu

SNAFU context selector for the Error::FetchTransaction variant

HeaderQueryData

InconsistentLeafError

InvalidTransactionIndexSnafu

SNAFU context selector for the Error::InvalidTransactionIndex variant

LeafQueryData

Limits

Options

PayloadMetadata

A summary of a payload without all the data.

PayloadQueryData

QuerySnafu

SNAFU context selector for the Error::Query variant

RangeLimitSnafu

SNAFU context selector for the Error::RangeLimit variant

RequestSnafu

SNAFU context selector for the Error::Request variant

TransactionQueryData

TransactionSummaryQueryData

VidCommonMetadata

A summary of a VID payload without all the data.

VidCommonQueryData

Enums

BlockId

Error

Fetch

An in-progress request to fetch some data.

LeafId

Traits

AvailabilityDataSource

An interface for querying a HotShot blockchain.

QueryableHeader

QueryablePayload

A block payload whose contents (e.g. individual transactions) can be examined.

UpdateAvailabilityData

Functions

define_api

enforce_range_limit 🔒

Type Aliases

BlockHash

A block hash is the hash of the block header.

FetchStream

LeafHash

QcHash

Timestamp

TransactionHash

TransactionInclusionProof

TransactionIndex