Module hotshot_query_service::data_source

Persistent storage and sources of data consumed by APIs.

The APIs provided by this query service are generic over the implementation which actually retrieves data in answer to queries. We call this implementation a data source. This module defines a data source and provides several pre-built implementations:

FileSystemDataSource

SqlDataSource

FetchingDataSource, a generalization of the above

MetricsDataSource

The user can choose which data source to use when initializing the query service.

We also provide combinators for modularly adding functionality to existing data sources:

ExtensibleDataSource

Re-exports

pub use fetching::AvailabilityProvider;

pub use fetching::FetchingDataSource;

pub use sql::SqlDataSource;

Modules

availability_tests

Generic tests we can instantiate for all the availability data sources.

fetching

Asynchronous retrieval of missing data.

node_tests

Generic tests we can instantiate for all the node data sources.

persistence_tests

Generic tests we can instantiate for any data source with reliable, versioned persistent storage.

sql

status_tests

Generic tests we can instantiate for all the status data sources.

storage

Persistent storage for data sources.

Structs

ExtensibleDataSource

Wrapper to add extensibility to an existing data source.

MetricsDataSource

A minimal data source for the status API provided in this crate, with no persistent storage.

Traits

UpdateDataSource

An extension trait for types which implement the update trait for each API module.

VersionedDataSource

A data source with an atomic transaction-based synchronization interface.

Type Aliases

FileSystemDataSource

A data source for the APIs provided in this crate, backed by the local file system.