hotshot_query_service::data_source
Module sql

Re-exports

pub use sql::Transaction;

Modules

testing

Macros

include_migrations

Embed migrations from the given directory into the current binary for PostgreSQL or SQLite.

Structs

Error

The Error type, a wrapper around a dynamic error type.

Migration

Represents a schema migration to be run on the database, this struct is used by the embed_migrations! macro to gather migration files and shouldn’t be needed by the user

Type Aliases

Builder

Config

SqlDataSource

A data source for the APIs provided in this crate, backed by a remote PostgreSQL database.