Module hotshot_query_service::data_source::storage::sql

Re-exports

pub use tokio_postgres as postgres;

Modules

testing

Macros

include_dir

Embed the contents of a directory in your crate.

include_migrations

Embed migrations from the given directory into the current binary.

Structs

Config

Postgres client config.

Error

The Error type, a wrapper around a dynamic error type.

Migration

Represents a schema migration to be run on the database, this struct is used by the embed_migrations! macro to gather migration files and shouldn’t be needed by the user

SqlStorage

Storage for the APIs provided in this crate, backed by a remote PostgreSQL database.

Transaction

An atomic SQL transaction.

Traits

Query

Functions

default_migrations

The migrations requied to build the default schema for this version of SqlStorage.