Trait hotshot_query_service::data_source::Transaction

pub trait Transaction: Send + Sync {
    // Required methods
    fn commit(self) -> impl Future<Output = Result<()>> + Send;
    fn revert(self) -> impl Future + Send;
}

A unit of atomicity for updating a shared data sourec.

The methods provided by this trait can be used to write such pending changes back to persistent storage (commit) so that they become visible to other clients of the same underlying storage, and are saved if the process restarts. It also allows pending changes to be rolled back (revert) so that they are never written back to storage and are no longer reflected even through the data source object which was used to make the changes.

Required Methods

fn commit(self) -> impl Future<Output = Result<()>> + Send

fn revert(self) -> impl Future + Send

Object Safety

This trait is not object safe.