Trait hotshot_query_service::data_source::UpdateDataSource

pub trait UpdateDataSource<Types: NodeType>: UpdateAvailabilityData<Types> + UpdateStatusData {
    // Required method
    fn update<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: &'life1 Event<Types>,
    ) -> Pin<Box<dyn Future<Output = Result<(), <Self as UpdateAvailabilityData<Types>>::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

An extension trait for types which implement the update trait for each API module.

If a type implements both UpdateAvailabilityData and UpdateStatusData, then it can be fully kept up to date through two interfaces:

populate_metrics, to get a handle for populating the status metrics, which should be used when initializing a SystemContextHandle

update, to update the query state when a new HotShot event is emitted

Required Methods

fn update<'life0, 'life1, 'async_trait>( &'life0 mut self, event: &'life1 Event<Types>, ) -> Pin<Box<dyn Future<Output = Result<(), <Self as UpdateAvailabilityData<Types>>::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Update query state based on a new consensus event.

The caller is responsible for authenticating event. This function does not perform any authentication, and if given an invalid event (one which does not follow from the latest known state of the ledger) it may panic or silently accept the invalid event. This allows the best possible performance in the case where the query service and the HotShot instance are running in the same process (and thus the event stream, directly from HotShot) is trusted.

If you want to update the data source with an untrusted event, for example one received from a peer over the network, you must authenticate it first.