Trait hotshot_query_service::data_source::VersionedDataSource

pub trait VersionedDataSource {
    type Error: Error + Debug + Send + Sync + 'static;

    // Required methods
    fn commit<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn revert<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A data source with an atomic transaction-based synchronization interface.

Any changes made to a versioned data source are initially visible when queried through that same data source object only. They are not immediately written back to storage, which means that a new data source object opened against the same persistent storage will not reflect the changes. In particular, this means that if the process restarts and reopens its storage, uncommitted changes will be lost.

The methods provided by this trait can be used to write such pending changes back to persistent storage (commit) so that they become visible to other clients of the same underlying storage, and are saved if the process restarts. It also allows pending changes to be rolled back (revert) so that they are never written back to storage and are no longer reflected even through the data source object which was used to make the changes.

Required Associated Types

type Error: Error + Debug + Send + Sync + 'static

Required Methods

fn commit<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Atomically commit to all outstanding modifications to the data.

If this method fails, outstanding changes are left unmodified. The caller may opt to retry or to erase outstanding changes with revert.

fn revert<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Erase all oustanding modifications to the data.

This function must not return if it has failed to revert changes. Inability to revert changes to the database is considered a fatal error, and this function may panic.