hotshot_query_service::explorer

Trait ExplorerTransaction

pub trait ExplorerTransaction<Types>
where
    Types: NodeType,
    Header<Types>: QueryableHeader<Types>,
{
    // Required methods
    fn namespace_id(&self) -> NamespaceId<Types>;
    fn payload_size(&self) -> u64;
}

ExplorerTransaction is a trait that allows the Explorer API to be able to retrieve a namespace id from a transaction. This is necessary for the Explorer API to be able to display the namespace id for a TransactionSummary.

Required Methods

fn namespace_id(&self) -> NamespaceId<Types>

namespace_id returns the namespace id of the individual transaction.

fn payload_size(&self) -> u64

payload_size returns the size of the payload of the transaction.