Module hotshot_query_service::fetching

Fetching missing data from remote providers.

This module provides a mechanism to fetch data that is missing from this query service’s storage from a remote data availability provider. Fetcher can be used to handle concurrent requests for data, ensuring that each distinct resource is only fetched once at a time.

Fetching is ultimately dispatched to a Provider, which implements fetching for a specific type of resource from a specific source. The provider module contains built-in implementations of Provider for various data availability sources.

Re-exports

pub use provider::Provider;

pub use request::Request;

Modules

provider

Asynchronous fetching from external data availability providers.

request

Requests for fetching resources.

Structs

Fetcher

Management of concurrent requests to fetch resources.

Traits

Callback

A callback to process the result of a request.

LocalCallback

A callback to process the result of a request.