hotshot_query_service
Module node

A node’s view of a HotShot chain

The node API provides a subjective view of the HotShot blockchain, from the perspective of one particular node. It provides access to information that the availability API does not, because this information depends on the perspective of the node observing it, and may be subject to eventual consistency. For example, /node/block-height may return smaller counts than expected, if the node being queried is not fully synced with the entire history of the chain. However, the node will eventually sync and return the expected counts.

Re-exports

pub use crate::availability::BlockHash;

pub use crate::availability::BlockId;

Modules

data_source 🔒

Data for the node API.

query_data 🔒

Structs

CustomSnafu

SNAFU context selector for the Error::Custom variant

Limits

Options

QuerySnafu

SNAFU context selector for the Error::Query variant

QueryVidSnafu

SNAFU context selector for the Error::QueryVid variant

QueryWindowSnafu

SNAFU context selector for the Error::QueryWindow variant

RequestSnafu

SNAFU context selector for the Error::Request variant

SyncStatus

TimeWindowQueryData

Response to a /:resource/window query.

Enums

Error

WindowStart

Traits

NodeDataSource

Functions

define_api